Moved tests to separate file + added filepaths to plotgenerator.py
Made the change to the plotgenerator.py file that allows it to use the
file path from the settings file. NOTE: Due to my previous issues with
installing jaratoolbox, I can’t test this from the folder in Git (since
I have to be in my root in order to import the toolbox). In theory, it
should work fine. However, if issues with the code arises let me know
so I can attempt to fix them.